feat: modularize components in the table
- Extract `RepositoryCell` and `RepositoryActionCell` components to
improve structure and readability.
- Add reusable `useRowActions` hook to centralize row action logic.
- Replace inline styles with PatternFly utilities.
- Refine filter and tooltip implementations.
